+++ sdk/servicebus/azure-servicebus/samples/sync_samples/receive_deadlettered_messages.py
@@ -22,7 +22,7 @@ servicebus_client = ServiceBusClient.from_connection_string(conn_str=CONNECTION_
 with servicebus_client:
     receiver = servicebus_client.get_queue_deadletter_receiver(queue_name=QUEUE_NAME, prefetch=10)
     with receiver:
-        received_msgs = receiver.receive(max_batch_size=10, max_wait_time=5)
+        received_msgs = receiver.receive_messages(max_batch_size=10, max_wait_time=5)
         for msg in received_msgs:
             print(str(msg))
             msg.complete()