+++ sdk/servicebus/azure-servicebus/samples/sync_samples/receive_subscription.py
@@ -27,7 +27,7 @@ with servicebus_client:
         prefetch=10
     )
     with receiver:
-        received_msgs = receiver.receive(max_batch_size=10, max_wait_time=5)
+        received_msgs = receiver.receive_messages(max_batch_size=10, max_wait_time=5)
         for msg in received_msgs:
             print(str(msg))
             msg.complete()