+++ sdk/servicebus/azure-servicebus/samples/sync_samples/session_send_receive.py
@@ -41,7 +41,7 @@ def receive_batch_message(receiver):
     session = receiver.session
     session.set_session_state("START")
     print("Session state:", session.get_session_state())
-    received_msgs = receiver.receive(max_batch_size=10, max_wait_time=5)
+    received_msgs = receiver.receive_messages(max_batch_size=10, max_wait_time=5)
     for msg in received_msgs:
         print(str(msg))
         msg.complete()