+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_queues_async.py
@@ -105,7 +105,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 for i in range(5):
                     await sender.send(Message("Message {}".format(i)))
             async with sb_client.get_queue_receiver(servicebus_queue.name, mode=ReceiveSettleMode.ReceiveAndDelete, idle_timeout=5) as messages:
-                batch = await messages.receive()
+                batch = await messages.receive_messages()
                 count = len(batch)
                 async for message in messages:
                    _logger.debug(message)
@@ -495,13 +495,13 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                         await sender.send(message)
 
                 count = 0
-                messages = await receiver.receive()
+                messages = await receiver.receive_messages()
                 while messages:
                     for message in messages:
                         print_message(_logger, message)
                         count += 1
                         await message.dead_letter(reason="Testing reason", description="Testing description")
-                    messages = await receiver.receive()
+                    messages = await receiver.receive_messages()
 
             assert count == 10
 
@@ -542,13 +542,13 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                         await sender.send(message)
 
                 count = 0
-                messages = await receiver.receive()
+                messages = await receiver.receive_messages()
                 while messages:
                     for message in messages:
                         print_message(_logger, message)
                         await message.dead_letter(reason="Testing reason", description="Testing description")
                         count += 1
-                    messages = await receiver.receive()
+                    messages = await receiver.receive_messages()
 
             assert count == 10
 
@@ -596,7 +596,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     await sender.send(message)
 
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.peek(5)
+                messages = await receiver.peek_messages(5)
                 assert len(messages) == 5
                 assert all(isinstance(m, PeekMessage) for m in messages)
                 for message in messages:
@@ -619,7 +619,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                         message = Message("Test message no. {}".format(i))
                         await sender.send(message)
 
-                messages = await receiver.peek(5)
+                messages = await receiver.peek_messages(5)
                 assert len(messages) > 0
                 assert all(isinstance(m, PeekMessage) for m in messages)
                 for message in messages:
@@ -637,7 +637,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
 
             async with sb_client.get_queue_receiver(servicebus_queue.name, idle_timeout=5, mode=ReceiveSettleMode.PeekLock, prefetch=10) as receiver:
-                messages = await receiver.peek(10)
+                messages = await receiver.peek_messages(10)
                 assert len(messages) == 0
 
     @pytest.mark.liveTest
@@ -657,10 +657,10 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                         message = Message("Test message no. {}".format(i))
                         await sender.send(message)
 
-                messages.extend(await receiver.receive())
+                messages.extend(await receiver.receive_messages())
                 recv = True
                 while recv:
-                    recv = await receiver.receive()
+                    recv = await receiver.receive_messages()
                     messages.extend(recv)
 
                 try:
@@ -764,7 +764,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
 
             time.sleep(30)
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
             assert not messages
 
             async with sb_client.get_queue_deadletter_receiver(servicebus_queue.name, idle_timeout=5, mode=ReceiveSettleMode.PeekLock) as receiver:
@@ -816,7 +816,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await sender.send(message)
 
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
 
             with pytest.raises(MessageSettleFailed):
@@ -837,7 +837,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await sender.send(message)
 
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
                 time.sleep(60)
                 assert messages[0].expired
@@ -847,7 +847,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     await messages[0].renew_lock()
 
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=30)
+                messages = await receiver.receive_messages(max_wait_time=30)
                 assert len(messages) == 1
                 print_message(_logger, messages[0])
                 assert messages[0].header.delivery_count > 0
@@ -868,7 +868,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await sender.send(message)
             
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
                 time.sleep(15)
                 await messages[0].renew_lock()
@@ -879,7 +879,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await messages[0].complete()
             
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 0
 
     @pytest.mark.liveTest
@@ -896,7 +896,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await sender.send(message)
 
             async with sb_client.get_queue_receiver(servicebus_queue.name, mode=ReceiveSettleMode.ReceiveAndDelete) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
                 received = messages[0]
                 print_message(_logger, received)
@@ -913,7 +913,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
 
             time.sleep(30)
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 for m in messages:
                     print_message(_logger, m)
                 assert len(messages) == 0
@@ -934,10 +934,10 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await sender.send(message)
 
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=10)
+                messages = await receiver.receive_messages(max_wait_time=10)
                 recv = True
                 while recv:
-                    recv = await receiver.receive(max_wait_time=10)
+                    recv = await receiver.receive_messages(max_wait_time=10)
                     messages.extend(recv)
 
                 assert len(messages) == 5
@@ -964,7 +964,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     message.scheduled_enqueue_time_utc = enqueue_time
                     await sender.send(message)
 
-                messages = await receiver.receive(max_wait_time=120)
+                messages = await receiver.receive_messages(max_wait_time=120)
                 if messages:
                     try:
                         data = str(messages[0])
@@ -1001,9 +1001,9 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                     tokens = await sender.schedule([message_a, message_b], enqueue_time)
                     assert len(tokens) == 2
 
-                recv = await receiver.receive(max_wait_time=120)
+                recv = await receiver.receive_messages(max_wait_time=120)
                 messages.extend(recv)
-                recv = await receiver.receive(max_wait_time=5)
+                recv = await receiver.receive_messages(max_wait_time=5)
                 messages.extend(recv)
                 if messages:
                     try:
@@ -1038,7 +1038,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
 
                     await sender.cancel_scheduled_messages(tokens)
 
-                messages = await receiver.receive(max_wait_time=120)
+                messages = await receiver.receive_messages(max_wait_time=120)
                 assert len(messages) == 0
 
     @pytest.mark.liveTest
@@ -1059,7 +1059,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
 
             async with sb_client.get_queue_receiver(servicebus_queue.name, mode=ReceiveSettleMode.ReceiveAndDelete) as receiver:
                 assert receiver._config.transport_type == TransportType.AmqpOverWebsocket
-                messages = await receiver.receive(max_wait_time=5)
+                messages = await receiver.receive_messages(max_wait_time=5)
                 assert len(messages) == 1
 
     def test_queue_message_http_proxy_setting(self):
@@ -1098,7 +1098,7 @@ class ServiceBusQueueAsyncTests(AzureMgmtTestCase):
                 await sender.send(message)
 
             async with sb_client.get_queue_receiver(servicebus_queue.name) as receiver:
-                messages = await receiver.receive(max_wait_time=5)
+                messages = await receiver.receive_messages(max_wait_time=5)
                 await receiver._handler.message_handler.destroy_async()  # destroy the underlying receiver link
                 assert len(messages) == 1
                 await messages[0].complete()