+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Fixed bug where sync AutoLockRenew does not shutdown itself timely.
 * Fixed bug where async AutoLockRenew does not support context manager.
+* Renamed `Receive()` and `Peek()` to `Receive_Messages()` and `Peek_Messages()` to align with other service bus SDKs.
 
 ## 7.0.0b3 (2020-06-08)